requirements.txt
flask
ocpp
websockets
requests
pyserial
